Chain Update
Redefined the AudioChain data structure, how it interacts as a
linked-list with forking and merging capabilities. The fork and merge
functionality has been outlined, but is not yet implemented in the
class. The class is being pushed at this point so that I can test some
functionality on the Ubuntu VM.